llm/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic==2.4.2
httpx>=0.25.0
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
accelerate==0.27.2
bitsandbytes==0.42.0
huggingface-hub>=0.19.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
sentencepiece==0.1.99
protobuf==3.20.3
einops==0.7.0 